pc_store_management/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
mysql-connector-python

